Modify dgraphloader to be able to load geo files
